+++ internal/provider/provider_sdk.go
@@ -62,7 +62,6 @@ func Provider(v string) *schema.Provider {
 			"scalr_account_allowed_ips":            resourceScalrAccountAllowedIps(),
 			"scalr_agent_pool":                     resourceScalrAgentPool(),
 			"scalr_module":                         resourceScalrModule(),
-			"scalr_policy_group":                   resourceScalrPolicyGroup(),
 			"scalr_policy_group_linkage":           resourceScalrPolicyGroupLinkage(),
 			"scalr_provider_configuration":         resourceScalrProviderConfiguration(),
 			"scalr_provider_configuration_default": resourceScalrProviderConfigurationDefault(),